+++ src/NullType.php
@@ -16,6 +16,16 @@ final class NullType extends Type
         return !($other instanceof VoidType);
     }
 
+    public function name(): string
+    {
+        return 'null';
+    }
+
+    public function asString(): string
+    {
+        return 'null';
+    }
+
     public function getReturnTypeDeclaration(): string
     {
         return '';