+++ src/UnknownType.php
@@ -16,6 +16,16 @@ final class UnknownType extends Type
         return true;
     }
 
+    public function name(): string
+    {
+        return 'unknown type';
+    }
+
+    public function asString(): string
+    {
+        return '';
+    }
+
     public function getReturnTypeDeclaration(): string
     {
         return '';