+++ tests/unit/CallableTypeTest.php
@@ -16,9 +16,9 @@ use SebastianBergmann\Type\TestFixture\ClassWithInvokeMethod;
 /**
  * @covers \SebastianBergmann\Type\CallableType
  *
- * @uses \SebastianBergmann\Type\Type
  * @uses \SebastianBergmann\Type\ObjectType
  * @uses \SebastianBergmann\Type\SimpleType
+ * @uses \SebastianBergmann\Type\Type
  * @uses \SebastianBergmann\Type\TypeName
  */
 final class CallableTypeTest extends TestCase
@@ -142,7 +142,8 @@ final class CallableTypeTest extends TestCase
     {
         $this->assertFalse(
             $this->type->isAssignable(
-                Type::fromValue(new class {
+                Type::fromValue(new class
+                {
                 }, false)
             )
         );