+++ tests/unit/TypeTest.php
@@ -14,12 +14,12 @@ use PHPUnit\Framework\TestCase;
 /**
  * @covers \SebastianBergmann\Type\Type
  *
- * @uses \SebastianBergmann\Type\SimpleType
- * @uses \SebastianBergmann\Type\GenericObjectType
- * @uses \SebastianBergmann\Type\ObjectType
- * @uses \SebastianBergmann\Type\TypeName
  * @uses \SebastianBergmann\Type\CallableType
+ * @uses \SebastianBergmann\Type\GenericObjectType
  * @uses \SebastianBergmann\Type\IterableType
+ * @uses \SebastianBergmann\Type\ObjectType
+ * @uses \SebastianBergmann\Type\SimpleType
+ * @uses \SebastianBergmann\Type\TypeName
  */
 final class TypeTest extends TestCase
 {